requirements.txt
django==2.2.4
djangorestframework==3.10.3
django-extensions==2.2.5
celery==4.3.0
redis==3.3.11
psycopg2-binary==2.8.3
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl==7.1.0
rest-framework-generic-relations==1.2.2
drf-nested-routers==0.91
